python/requirements.txt
nats-py
protobuf>=6.32.1